Use size_t for count variables

+++ IRasterizer.h
@@ -20,9 +20,9 @@ public:
     virtual void drawLine(const Vertex &v0, const Vertex &v1) const = 0;
     virtual void drawTriangle(const Vertex& v0, const Vertex &v1, const Vertex &v2) const = 0;
 
-    virtual void drawPointList(const Vertex *vertices, const int *indices, int indexCount) const = 0;
-    virtual void drawLineList(const Vertex *vertices, const int *indices, int indexCount) const = 0;
-	virtual void drawTriangleList(const Vertex *vertices, const int *indices, int indexCount) const = 0;
+    virtual void drawPointList(const Vertex *vertices, const int *indices, size_t indexCount) const = 0;
+    virtual void drawLineList(const Vertex *vertices, const int *indices, size_t indexCount) const = 0;
+	virtual void drawTriangleList(const Vertex *vertices, const int *indices, size_t indexCount) const = 0;
 };
 
 } // end namespace swr
